feat(settings): Add categorized settings navigation
- Split the monolithic settings page into focused section components
- Add accent color and background glow customization with live previews
- Add transparent sidebar preference and persist new appearance settings
- Add navigation labels and translations for the new settings structure
