+++ electron/main/services/settings.ts
@@ -42,6 +42,8 @@ const DEFAULT_SETTINGS: AppSettings = {
     discordRpcEnabled: false,
     contributeMatchSalts: false,
     unifiedLaunchButton: false,
+    sidebarTransparent: false,
+    backgroundGradient: null,
     verboseModTrace: false,
 };
