+++ src/locales/manifest.json
@@ -1,24 +1,24 @@
 {
   "sourceLanguage": "en",
-  "totalKeys": 2051,
+  "totalKeys": 2070,
   "languages": [
     {
       "code": "bg",
       "name": "български",
       "translatedKeys": 1964,
-      "pct": 96
+      "pct": 95
     },
     {
       "code": "en",
       "name": "English",
-      "translatedKeys": 2051,
+      "translatedKeys": 2070,
       "pct": 100
     },
     {
       "code": "fr",
       "name": "français",
       "translatedKeys": 1858,
-      "pct": 91
+      "pct": 90
     },
     {
       "code": "ru",
